Add solution for Day 10 Part 2
This one required more than just coding the problem, since brute
force would have been prohibitive.  The key insight is that the
only places that can change are strings of 1 different adaptors.
So given the string of "differences" that was easily generated
for Part 1, we just needed a function to generate a list of the
lengths of the 1-diff runs that are separated by 3-diffs.  After
a simple conversion from length to possible combinations (which
was complicated at first by my failure to realize that removing
1-diffs from a string didn't have to be contiguous), we can
just multiply all those combos together to get the total number
of possiblities.